+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Change Log
 
 ## Unreleased
-- persons: country_short_blue, country_short_white を入れた方が良くないか？
+- map で db を置き換える。
+- 例外を出さない。upsert
 
-## 0.2.0-SNAPSHOT
+## 0.2.0 - 2022-08-13
 - src/fights/db.clj create tables. `persons`, `contests` and `competitions`.
 - test/fights/db_test.clj
 - src/fights/competitor.clj
@@ -11,6 +12,7 @@
 - (insert-person-id id_competition)
 - (insert-pserons year)
 - transaction でスピードアップできないか？
+- persons: country_short_blue, country_short_white を入れた。
 
 
 ## 0.1.0 - 2022-08-12